refactor: Replace FileUtil with fs-extra (#5109)
